move the profile to pages folder relates #9
