+++ client/src/App/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import NavBar from '../Components/Layout/Navbar';
 import Footer from '../Components/Layout/Footer';
-import KindergartenProfilePage from '../containers/KindergartenProfilePage';
+import KindergartenProfilePage from '../Pages/KindergartenProfilePage';
 
 import Home from '../Pages/HomePage';
 import './style.css';
