+++ client/src/Pages/KindergartenProfilePage/index.js
